% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barroso.R
\name{barroso}
\alias{barroso}
\title{Metrics from Barroso et al.}
\usage{
barroso(
  lightVar,
  sampling_int,
  unit_out_clength,
  loop = FALSE,
  na_rm = TRUE,
  as_df = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{sampling_int}{Numeric. Sampling interval in seconds.}

\item{unit_out_clength}{String. Time unit of the bright/dark cluster length.
Possible values are ("seconds", "minutes", "hours", "days"), which can be
abbreviated.}

\item{loop}{Logical. Should the data be looped? Defaults to FALSE.}

\item{na_rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Data frame or numeric.
}
\description{
Metrics from Barroso et al.
}
