% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bright_dark_period.R
\name{bright_dark_period}
\alias{bright_dark_period}
\title{Brightest or darkest continuous period}
\usage{
bright_dark_period(
  lightVar,
  timeVar,
  period_type,
  timespan,
  sampling_int,
  loop = FALSE,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{timeVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{period_type}{String indicating the type of period. Must be "bright" or
"dark".}

\item{timespan}{Single string or vector of strings with the timespan(s).
Timespans must be in the format "[numeric] [unit]", with possible units
("seconds","minutes","hours","days"). Units can be abbreviated.
See \code{\link{parse_timeunit_tosecs}}.}

\item{sampling_int}{Numeric. Sampling interval in seconds.}

\item{loop}{Logical. Should the data be looped? Defaults to FALSE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of timespan and calculated values.
   If wide is TRUE then variable names will be concatenated with the timespan.
}
\description{
This function finds the brightest or darkest continuous period of a given
timespan and calculates its \code{mean} light level, \code{onset},
\code{midpoint}, and \code{offset}. Defined as the period with the maximum
or minimum mean light level.
}
\details{
Assumes regular 24h light data. Otherwise, results may not be
   meaningful. Looping the data is recommended for finding the darkest period.
   Missing light values will be removed by default.
}
