% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_max.R
\name{cat_max}
\alias{cat_max}
\title{Length of longest continuous cluster above/below threshold}
\usage{
cat_max(
  lightVar,
  threshold,
  sampling_int = NULL,
  unit_out = NULL,
  loop = FALSE,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{threshold}{Single numeric value or vector specifying threshold
intensities. The sign indicates above/below (see \code{\link{threshold}}).}

\item{sampling_int}{Numeric. Sampling interval in seconds. If not specified
(default), no time scaling will be performed.}

\item{unit_out}{Character. Time unit of output. Possible values are
("seconds", "minutes", "hours", "days"), which can be abbreviated.
If not specified (default), no time scaling will be performed.}

\item{loop}{Logical. Should the data be looped? Defaults to FALSE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of threshold and calculated values.
   If wide is TRUE then variable names will be concatenated with the threshold.
}
\description{
Length of longest continuous cluster above/below threshold
}
