% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_var.R
\name{coeff_var}
\alias{coeff_var}
\title{Coefficient of variation}
\usage{
coeff_var(lightVar, na_rm = TRUE, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{na_rm}{Logical. Should missing light values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Data frame or numeric vector.
}
\description{
This function calculates the coefficient of variation, defined as the standard
deviation divided by the mean.
}
