% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_exposure.R
\name{cumulative_exposure}
\alias{cumulative_exposure}
\title{Cumulative exposure (CE)}
\usage{
cumulative_exposure(lightVar, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
This function calculates the integral of light exposure across the given
interval.
}
