% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_index.R
\name{disparity_index}
\alias{disparity_index}
\title{Disparity index}
\usage{
disparity_index(lightVar, na_rm = TRUE, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{na_rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Data frame or numeric vector.
}
\description{
This function calculates the continuous disparity index as described in
Fernández-Martínez et al. (2018).
}
\references{
Fernández-Martínez, M., Vicca, S., Janssens, I. A., Carnicer, J.,
  Martín-Vide, J., & Peñuelas, J. (2018).
  The consecutive disparity index, D: A measure of temporal variability in
  ecological studies. \emph{Ecosphere}, 9(12), e02527.
  \url{https://doi.org/10.1002/ecs2.2527}
}
