% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_tatr.R
\name{dose_tatr}
\alias{dose_tatr}
\title{Dose within threshold range}
\usage{
dose_tatr(
  lightVar,
  lower,
  upper,
  sampling_int = NULL,
  unit_out = NULL,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{lower}{Single numeric value or vector specifying the lower bound of the
threshold intensity range. The sign indicates above/below
(see \code{\link{threshold}}). Must be same length as upper bound.}

\item{upper}{Single numeric value or vector specifying the upper bound of the
threshold intensity range. The sign indicates above/below
(see \code{\link{threshold}}). Must be same length as lower bound.}

\item{sampling_int}{Numeric. Sampling interval in seconds. If not specified
(default), no time scaling will be performed.}

\item{unit_out}{Character. Time unit of output. Possible values are
("seconds", "minutes", "hours", "days"), which can be abbreviated.
If not specified (default), no time scaling will be performed.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of threshold and calculated values.
   If `wide` is TRUE then variable names will be concatenated with the
   threshold.
}
\description{
This function calculates the light dose within a specified intensity range.
}
