% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_light}
\alias{example_light}
\title{Example light data.}
\format{
A data frame with 1440 rows and 3 variables:
\describe{
  \item{datetime}{datetime, as yy/mm/dd HH:MM:SS}
  \item{light}{illuminance, in lux}
  \item{awake}{logical, whether subject was awake}
}
}
\usage{
example_light
}
\description{
A dataset containing personal light exposure data measured across 24 hours
with a recording interval of 60 seconds
}
\keyword{datasets}
