% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{fic}
\alias{fic}
\title{Frequency of intensity changes}
\usage{
fic(lightVar, threshold, na_rm = FALSE, as_df = TRUE, wide = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{threshold}{Single numeric value or vector specifying threshold
intensities. The sign indicates above/below (see \code{\link{threshold}}).}

\item{na_rm}{Logical. Should missing light values be removed? Defaults to FALSE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of threshold and calculated values.
   If `wide` is TRUE then variable names will be concatenated with the threshold.
}
\description{
This functions calculates the times a given threshold
intensity is crossed.
}
\references{
Alvarez, A. A., & Wildsoet, C. F. (2013). Quantifying light
   exposure patterns in young adult students. \emph{Journal of Modern Optics},
   60(14), 1200–1208. \url{https://doi.org/10.1080/09500340.2013.845700}
}
