% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_to_rows.R
\name{interval_to_rows}
\alias{interval_to_rows}
\title{Interval to data frame rows}
\usage{
interval_to_rows(df, rowvar, from, to, by, removeBounds = TRUE)
}
\arguments{
\item{df}{The data frame to be reshaped.}

\item{rowvar}{Name of the new variable holding the row indices.}

\item{from}{Start of interval. Numeric or POSIXct.}

\item{to}{End of interval. Numeric or POSIXct.}

\item{by}{Numeric. Spacing between individual rows. Needs to be in same unit
as interval variables (e.g., seconds for POSIXct)}

\item{removeBounds}{Logical. Should start and end variables be removed?
Defaults to TRUE.}
}
\value{
Reshaped data frame.
}
\description{
This function reshapes a data frame from a given interval to continuously
spaced rows.
}
