% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intradaily_variability.R
\name{intradaily_variability}
\alias{intradaily_variability}
\title{Intradaily variability (IV)}
\usage{
intradaily_variability(lightVar, dtVar, na.rm = TRUE, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{dtVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
This function calculates the variability of consecutive light levels within
a 24h day. Calculated as the ratio of the variance of the differences between
consecutive light levels to the total variance across the day. Calculated for
mean hourly light levels. Higher values indicate more fragmentation.
}
\references{
Van Someren, E. J. W., Swaab, D. F., Colenda, C. C., Cohen, W.,
   McCall, W. V., & Rosenquist, P. B. (1999). Bright Light Therapy: Improved
   Sensitivity to Its Effects on Rest-Activity Rhythms in Alzheimer Patients
   by Application of Nonparametric Methods. \emph{Chronobiology International},
   16(4), 505–518. \url{https://doi.org/10.3109/07420529908998724}
}
