% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funs.R
\name{math_funs}
\alias{math_funs}
\alias{geomean}
\alias{geosd}
\alias{se}
\alias{logp1}
\title{Mathematical helper functions}
\usage{
geomean(x, trim = 0, na.rm = FALSE, base = exp(1))

geosd(x, na.rm = FALSE, base = exp(1))

se(x, na.rm = FALSE)

logp1(x, base = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{trim}{The fraction (0 to 0.5) of observations to be trimmed from each
end of x before the mean is computed. Values of trim outside that range
are taken as the nearest endpoint. See \code{\link[base]{mean}}}

\item{na.rm}{Logical. Should missing values be removed?}

\item{base}{A positive or complex number: the base with respect to which
logarithms are computed. Defaults to `exp(1)`.}
}
\description{
Mathematical helper functions
}
\details{
`geomean` calculates the geometric mean.

`geosd` calculates the geometric standard deviation

`se` calculates the standard error

`logp1` calculates log(x + 1) for any base. The default base is 10.
}
