% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_period.R
\name{max_min_period}
\alias{max_min_period}
\title{Maximum or minimum continuous period}
\usage{
max_min_period(
  lightVar,
  dtVar,
  timespan,
  period_type,
  sampling_int = 60,
  loop = TRUE,
  na.rm = TRUE,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{dtVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{timespan}{Single string or vector of strings with the timespan(s).
Timespans must be in the format "[numeric] [unit]", with possible units
("seconds","minutes","hours","days"). Units can be abbreviated.
See \code{\link{parse_timeunit_tosecs}}.}

\item{period_type}{String indicating the type of period. Must be "max" or "min".}

\item{sampling_int}{Numeric. Sampling interval in seconds. Defaults to 60.}

\item{loop}{Logical. Should the data be looped? Defaults to TRUE}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of timespan and calculated values.
   If wide is TRUE then variable names will be concatenated with the timespan.
}
\description{
This function finds the brightest or darkest continuous period of a given
timespan and calculates its `mean` light level, `onset`, `midpoint`, and
`offset`. Defined as the period with the maximum or minimum mean light level.
}
\details{
Assumes 24h light data. Otherwise, results may not be meaningful.
   Looping the data is recommended for finding the minimum (darkest) period.
}
