% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvR.R
\name{nvR_circadian_metrics}
\alias{nvR_circadian_metrics}
\alias{nvR_circadianDisturbance}
\alias{nvR_circadianBias}
\alias{nvR_relativeAmplitudeError}
\title{Performance metrics for circadian response}
\usage{
nvR_circadianDisturbance(RC, RC_ref, as_df = TRUE)

nvR_circadianBias(RC, RC_ref, as_df = TRUE)

nvR_relativeAmplitudeError(RC, RC_ref, as_df = TRUE)
}
\arguments{
\item{RC}{Time series of circadian response (see \code{\link{circadianResponse}}.
Length must be 24h.}

\item{RC_ref}{Time series of circadian response (see \code{\link{circadianResponse}}
for reference light exposure pattern. Length must be 24h.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\description{
Performance metrics for circadian response
}
\details{
`nvR_circadianDisturbance()` calculates the circadian disturbance (CD).

`nvR_circadianBias()` calculates the circadian bias (CB).

`nvR_relativeAmplitudeError()` calculates the relative amplitude error (RAE).
}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
   indicators of non-visual health potential in the built environment.
   [Doctoral dissertation, EPFL]. EPFL infoscience.
   \url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
