% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvR.R
\name{nvR_circadianResponse}
\alias{nvR_circadianResponse}
\title{Non-visual circadian response}
\usage{
nvR_circadianResponse(
  irr_mel,
  lux,
  sampling_int,
  datetime = NULL,
  sleep_onset = NULL,
  shiftModel = FALSE
)
}
\arguments{
\item{irr_mel}{Numeric vector. Melanopic irradiance values.}

\item{lux}{Numeric vector. Photopic illuminance values.}

\item{sampling_int}{Numeric. Sampling interval in seconds.}

\item{datetime}{POSIXct vector. Datetime values. Will only be used if
\code{sleep_onset} is specified. Default is NULL.}

\item{sleep_onset}{Character. Time of habitual sleep onset in the format
"HH:MM:SS". Will only be used if \code{datetime} is specified. Default is
NULL.}

\item{shiftModel}{Logical. Should the spectral shift model be used? Default
is FALSE.}
}
\value{
Numeric vector with circadian response values.
}
\description{
This function calculates the non-visual circadian response
}
\details{
If \code{datetime} and \code{sleep_onset} are not specified, the
   start of the timeseries is assumed to be at habitual sleep onset. The
   timeseries is assumed to be regular. Missing values in the light data will
   be replaced by 0.
}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
   indicators of non-visual health potential in the built environment.
   [Doctoral dissertation, EPFL]. EPFL infoscience.
   \url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
