% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvR.R
\name{nvR_direct_metrics}
\alias{nvR_direct_metrics}
\alias{nvR_cumulativeResponse}
\title{Performance metrics for direct response}
\usage{
nvR_cumulativeResponse(RD, sampling_int, unit_out = "hours", as_df = TRUE)
}
\arguments{
\item{RD}{Time series of direct response (see \code{\link{directResponse}}.}

\item{sampling_int}{Numeric. Sampling interval in seconds.}

\item{unit_out}{Character. Time unit of output. Possible values are
("seconds", "minutes", "hours", "days"), which can be abbreviated.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
Performance metrics for direct response
}
\details{
`nvR_cumulativeResponse()` calculates the cumulative direct response.
}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
   indicators of non-visual health potential in the built environment.
   [Doctoral dissertation, EPFL]. EPFL infoscience.
   \url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
