% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvR.R
\name{nvR_directResponse}
\alias{nvR_directResponse}
\title{Non-visual direct response}
\usage{
nvR_directResponse(irr_mel, lux, sampling_int, shiftModel = FALSE)
}
\arguments{
\item{irr_mel}{Numeric vector. Melanopic irradiance values.}

\item{lux}{Numeric vector. Photopic illuminance values.}

\item{sampling_int}{Numeric. Sampling interval in seconds.}

\item{shiftModel}{Logical. Should the spectral shift model be used? Default
is FALSE.}
}
\value{
Numeric vector with direct response values.
}
\description{
Non-visual direct response
}
\details{
The timeseries is assumed to be regular. Missing values in the
   light data will be replaced by 0.
}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
   indicators of non-visual health potential in the built environment.
   [Doctoral dissertation, EPFL]. EPFL infoscience.
   \url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
