% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_timeunit_tosecs.R
\name{parse_timeunit_tosecs}
\alias{parse_timeunit_tosecs}
\title{Parse time and unit and convert to seconds}
\usage{
parse_timeunit_tosecs(input)
}
\arguments{
\item{input}{String to be parsed. Must be in the format "[numeric] [unit]",
with possible units ("seconds","minutes","hours","days"). Units can be
abbreviated.}
}
\value{
List with converted time in seconds, the input time and unit as strings.
}
\description{
This is a convenience function for parsing a time together with its unit given
as a string and converting it to seconds.
}
\examples{
parse_timeunit_tosecs("30 minutes")
parse_timeunit_tosecs("0.5 hrs")
parse_timeunit_tosecs("1 d")
parse_timeunit_tosecs("500 secs")
\dontrun{
parse_timeunit_tosecs("30min")
}
}
