% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{pfilter}
\alias{pfilter}
\title{Filter and print}
\usage{
pfilter(df, ...)
}
\arguments{
\item{df}{Data frame to be filtered.}

\item{...}{Filter conditions. See \code{\link[dplyr]{filter}}}
}
\value{
Filtered data frame
}
\description{
This function filters data (like dplyr's filter) and prints the number of
rows filtered by each condition.
}
