% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_datetime.R
\name{regular_datetime}
\alias{regular_datetime}
\title{Regularise datetime vector}
\usage{
regular_datetime(dtVar, sampling_int)
}
\arguments{
\item{dtVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{sampling_int}{Numeric. Sampling interval (in seconds if `dtVar` is
POSIXct).}
}
\value{
Regularised time data.
}
\description{
This function fills in gaps in a datetime vector and returns an equally spaced
datetime vector.
}
