% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularise_data.R
\name{regularise_data}
\alias{regularise_data}
\title{Regularise (date)time vector}
\usage{
regularise_data(df, timeVar, sampling_int, startVar = NULL, endVar = NULL)
}
\arguments{
\item{df}{(Grouped) data frame to be regularised.}

\item{timeVar}{Column containing the time data. Can be POSIXct or numeric.}

\item{sampling_int}{Numeric. Sampling interval (in seconds if `timeVar` is
POSIXct).}

\item{startVar}{Column indicating the start of the interval to be regularised.
Can be POSIXct or numeric. If not specified, the first element of the
 `timeVar` vector will be taken (default).}

\item{endVar}{Column indicating the end of the interval to be regularised.
Can be POSIXct or numeric. If not specified, the last element of the
 `timeVar` vector will be taken (default).}
}
\value{
Regularised time data.
}
\description{
This function fills in gaps in a data frame containing time series data.
}
