% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularise_time.R
\name{regularise_time}
\alias{regularise_time}
\title{Regularise datetime vector}
\usage{
regularise_time(timeVar, sampling_int)
}
\arguments{
\item{timeVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{sampling_int}{Numeric. Sampling interval (in seconds if `timeVar` is
POSIXct).}
}
\value{
Regularised time data.
}
\description{
This function fills in gaps in a datetime vector and returns an equally spaced
datetime vector.
}
