% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_at_timespan.R
\name{threshold_at_timespan}
\alias{threshold_at_timespan}
\title{Find threshold at timespan}
\usage{
threshold_at_timespan(
  lightVar,
  timespan,
  sampling_int = NULL,
  above = TRUE,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{timespan}{Single value or vector specifying the timespan(s).
Can be numeric or string in the format "[numeric] [unit]", with possible units ("seconds","minutes",
"hours","days"). Units can be abbreviated, see
\code{\link{parse_timeunit_tosecs}}. If numeric, no time scaling will be performed.}

\item{sampling_int}{Numeric. Sampling interval in seconds. If not specified
(default), no time scaling will be performed.}

\item{above}{Logical. Should the threshold of the timespan at which the light
is above the threshold be calculated? Defaults to TRUE. If FALSE, the
threshold of the timespan at which the light is below the threshold is calculated.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of timespans and calculated values.
   If wide is TRUE, then variable names will be concatenated with the timespan.
}
\description{
This function finds the threshold of light for a given timespan.
This function is the inverse of \code{\link{tat}}.
}
