% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_conv.R
\name{time_conv}
\alias{time_conv}
\alias{to.secs}
\alias{from.secs}
\alias{to.mins}
\alias{from.mins}
\alias{to.hours}
\alias{from.hours}
\alias{to.days}
\alias{from.days}
\title{Convert time to/from other unit.}
\usage{
to.secs(t, unit)

from.secs(t, unit)

to.mins(t, unit)

from.mins(t, unit)

to.hours(t, unit)

from.hours(t, unit)

to.days(t, unit)

from.days(t, unit)
}
\arguments{
\item{t}{Numeric time value.}

\item{unit}{Character time unit. Possible values are ("secs", "mins", "hours",
"days"). Can be abbreviated.}
}
\value{

}
\description{
Convenience functions for converting between time units.
}
