% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlit.R
\name{xlit}
\alias{xlit}
\alias{mlit}
\alias{flit}
\alias{llit}
\title{Mean/first/last timing above/below threshold.}
\usage{
mlit(lightVar, timeVar, threshold, as_df = TRUE, wide = TRUE)

flit(lightVar, timeVar, threshold, as_df = TRUE, wide = TRUE)

llit(lightVar, timeVar, threshold, as_df = TRUE, wide = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{timeVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{threshold}{Single numeric value or vector specifying threshold
intensities. The sign indicates above/below (see \code{\link{threshold}}).}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}
}
\value{
Data frame or matrix with pairs of threshold and calculated values.
   If wide is TRUE then variable names will be concatenated with the threshold.
}
\description{
These functions calculate the mean/first/last timepoint where light levels are
above or below a given threshold intensity within the given interval.
}
\details{
`mlit()` calculates the mean timing of light above/below threshold.

`flit()` calculates the first timing of light above/below threshold.

`llit()` calculates the last timing of light above/below threshold.
}
\references{
Reid, K. J., Santostasi, G., Baron, K. G., Wilson, J., Kang, J.,
   & Zee, P. C. (2014). Timing and Intensity of Light Correlate with Body Weight
    in Adults. \emph{PLOS ONE}, 9(4), e92251.
     \url{https://doi.org/10.1371/journal.pone.0092251}
}
