% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidLE.R
\name{centroidLE}
\alias{centroidLE}
\title{Centroid of light exposure}
\usage{
centroidLE(lightVar, dtVar, bin_size = NULL, na.rm = TRUE, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{dtVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{bin_size}{Numeric value or character specifying size of bins to average
light data. If dtVar is POSIXct it must be a string with the time followed
by its unit (e.g., "1 hour", "30 mins"). If NULL then no binning will be
performed. Defaults to NULL.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
This function calculates the centroid of light exposure as the mean of the
time vector weighted in proportion to the corresponding light intensity.
}
\references{
Phillips, A. J. K., Clerx, W. M., O’Brien, C. S., Sano, A., Barger,
   L. K., Picard, R. W., Lockley, S. W., Klerman, E. B., & Czeisler, C. A. (2017).
   Irregular sleep/wake patterns are associated with poorer academic performance
   and delayed circadian and sleep/wake timing. \emph{Scientific Reports},
   7(1), 3216. \url{https://doi.org/10.1038/s41598-017-03171-4}
}
