% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_exposure.R
\name{cumulative_exposure}
\alias{cumulative_exposure}
\title{Cumulative exposure (CE)}
\usage{
cumulative_exposure(lightVar, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
This function calculates the integral of light exposure across the given
interval.
}
\references{
Price, L. L. A. (2014). On the Role of Exponential Smoothing in
   Circadian Dosimetry. \emph{Photochemistry and Photobiology}, 90(5),
   \url{1184–1192. https://doi.org/10.1111/php.12282}
}
