% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpointCE.R
\name{midpointCE}
\alias{midpointCE}
\title{Midpoint of cumulative light exposure.}
\usage{
midpointCE(lightVar, dtVar, as_df = TRUE)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{dtVar}{Vector containing the time data. Can be POSIXct or numeric.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}
}
\value{
Single column data frame or vector.
}
\description{
This function calculates the timing corresponding to half of the cumulative
light exposure within the given time series.
}
