% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tatx.R
\name{tatx}
\alias{tatx}
\alias{tat}
\alias{tatr}
\title{Time above/below threshold or within threshold range}
\usage{
tat(
  lightVar,
  threshold,
  sampling_int = 60,
  unit_out = "mins",
  na.rm = TRUE,
  as_df = TRUE,
  wide = TRUE
)

tatr(
  lightVar,
  lower,
  upper,
  sampling_int = 60,
  unit_out = "mins",
  na.rm = TRUE,
  as_df = TRUE,
  wide = TRUE
)
}
\arguments{
\item{lightVar}{Numeric vector containing the light data.}

\item{threshold}{Single numeric value or vector specifying threshold
intensities. The sign indicates above/below (see \code{\link{threshold}}).}

\item{sampling_int}{Numeric. Sampling interval in seconds. Defaults to 60.}

\item{unit_out}{Character. Time unit of output. Possible values are
("secs", "mins", "hours", "days"). Can be abbreviated. Defaults to "mins".}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}

\item{as_df}{Logical. Should the output be returned as a data frame? Defaults
to TRUE.}

\item{wide}{Logical. Should the output be returned in wide format? Defaults to
TRUE.}

\item{lower}{Single numeric value or vector specifying the lower bound of the
threshold intensity range. The sign indicates above/below
(see \code{\link{threshold}}). Must be same length as upper bound.}

\item{upper}{Single numeric value or vector specifying the upper bound of the
threshold intensity range. The sign indicates above/below
(see \code{\link{threshold}}). Must be same length as lower bound.}
}
\value{
Data frame or matrix with pairs of threshold and calculated values.
   If wide is TRUE then variable names will be concatenated with the threshold.
}
\description{
These functions calculate the duration above/below a specified threshold
intensity or within a specified intensity range, within the time series provided.
}
\details{
`tat()` calculates the time above/below threshold.

`tatr()` calculates the time within threshold range.
}
